commit final avant restructuration de DB
