+++ class/Destination.php
@@ -7,8 +7,8 @@ class Destination
     private int $id;
     private string $location;
     private int $price;
-    private int $operatorId;
-    private string $img;
+    private int $tourOperatorId;
+    private string $imgDestination;
 
 
     public function __construct(array $data)
@@ -20,11 +20,19 @@ class Destination
     public function hydrate(array $data): void
     {
         foreach ($data as $key => $value) {
-            $method = 'set' . ucfirst($key);
+            $method = $this->transformToSetter($key);
             if (method_exists($this, $method)) {
                 $this->$method($value);
             }
         }
+    }
+
+
+    private function transformToSetter(string $arrayKey): string
+    {
+        $words = explode('_', $arrayKey);
+        $camelCaseName = implode('', array_map('ucfirst', $words));
+        return 'set' . $camelCaseName;
     }
 
 
@@ -64,26 +72,26 @@ class Destination
     }
 
 
-    public function getOperatorId(): int
+    public function getTourOperatorId(): int
     {
-        return $this->operatorId;
+        return $this->tourOperatorId;
     }
 
 
-    public function setOperatorId(int $OperatorId): void
+    public function setTourOperatorId(int $tourOperatorId): void
     {
-        $this->operatorId = $OperatorId;
+        $this->tourOperatorId = $tourOperatorId;
     }
 
 
-    public function getImg(): string
+    public function getImgDestination(): string
     {
-        return $this->img;
+        return $this->imgDestination;
     }
 
 
-    public function setImg(string $img): void
+    public function setImgDestination(string $imgDestination): void
     {
-        $this->img = $img;
+        $this->imgDestination = $imgDestination;
     }
 }
